profile-service/pyproject.toml
[project]
name = "profile-service"
version = "0.1.0"
description = ""
authors = [
    {name = "huyhuynh12hhk",email = "huyhuynh12.hh@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "django-storages (>=1.14.5,<2.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pynacl (>=1.5.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "drf-base64 (>=2.0,<3.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "kafka-python (>=2.0.4,<3.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "djangorestframework-camel-case (>=1.4.2,<2.0.0)",
    "drf-orjson-renderer (>=1.7.3,<2.0.0)",
    "python-keycloak (>=5.3.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
colorlog = "^6.9.0"
django-debug-toolbar = "^5.0.1"
pre-commit = "^4.1.0"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
pytest-xdist = "^3.6.1"
model-bakery = "^1.20.3"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "profile_service.core.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["cooking_core"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

